fixed eveyrthing needed for program to run and updated filing
