+++ src/main/java/Main.java
@@ -4,6 +4,14 @@ import database.CSVDatabase;
 import enums.WindowName;
 import ui.factories.*;
 import ui.windows.*;
+import ui.windows.creation.CreationDisplay;
+import ui.windows.game.GameController;
+import ui.windows.game.GameDisplay;
+import ui.windows.layout_managers.CardLayoutManager;
+import ui.windows.layout_managers.ICardLayoutManager;
+import ui.windows.layout_managers.PaneDelegator;
+import ui.windows.stats.StatsController;
+import ui.windows.stats.StatsDisplay;
 import use_cases.DataAccess;
 import use_cases.GameCreationInteractor;
 import use_cases.PlayerGameInteractor;
