+++ src/main/java/database/players.csv
@@ -1,0 +1,3 @@
+Solemann,5,6
+Lily,4,5
+Nancy,3,4
