+++ src/main/java/ui/windows/game/CardDelegator.java
@@ -1,4 +1,4 @@
-package ui.windows;
+package ui.windows.game;
 
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
