+++ src/main/java/ui/windows/game/GameDisplayData.java
@@ -1,4 +1,4 @@
-package ui.windows;
+package ui.windows.game;
 
 import java.util.ArrayList;
