+++ src/main/java/ui/windows/game/GameDrawDelegator.java
@@ -1,4 +1,6 @@
-package ui.windows;
+package ui.windows.game;
+
+import ui.windows.game.GameController;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
