+++ src/main/java/ui/windows/game/GameWindow.java
@@ -1,8 +1,11 @@
-package ui.windows;
+package ui.windows.game;
 
 import javax.swing.JPanel;
 
 import enums.WindowName;
+import ui.windows.Window;
+import ui.windows.game.GameDisplay;
+import ui.windows.layout_managers.PaneDelegator;
 
 public class GameWindow implements Window {
     private GameDisplay display;
