+++ src/main/java/ui/windows/layout_managers/PaneDelegator.java
@@ -1,4 +1,4 @@
-package ui.windows;
+package ui.windows.layout_managers;
 
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
