+++ src/main/java/ui/windows/menu/MenuDisplay.java
@@ -1,12 +1,12 @@
-package ui.windows;
+package ui.windows.menu;
 
 import java.awt.BorderLayout;
 
-import javax.swing.JButton;
 import javax.swing.JPanel;
 
 import enums.WindowName;
 import ui.components.NavigationButton;
+import ui.windows.layout_managers.PaneDelegator;
 
 public class MenuDisplay extends JPanel {
     private NavigationButton playButton;
