+++ src/main/java/ui/windows/rules/RulesDisplay.java
@@ -1,16 +1,16 @@
-package ui.windows;
+package ui.windows.rules;
 
 import java.awt.*;
 
 import javax.swing.BoxLayout;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
-import javax.swing.JButton;
 import javax.swing.JFrame;
 import javax.swing.border.EmptyBorder;
 
 import enums.WindowName;
 import ui.components.NavigationButton;
+import ui.windows.layout_managers.PaneDelegator;
 
 public class RulesDisplay extends JPanel {
     private NavigationButton backButton;
