+++ src/main/java/ui/windows/stats/StatsDisplay.java
@@ -1,7 +1,8 @@
-package ui.windows;
+package ui.windows.stats;
 
 import enums.WindowName;
 import ui.components.NavigationButton;
+import ui.windows.layout_managers.PaneDelegator;
 
 import javax.swing.*;
